requirements.txt
mcp[cli]
python-dotenv
fastapi
uvicorn
